Many minor style and bug fixes: - Remove spurious backslash in serve.cc - Use uint16_t for ports instead of int - Remove spurious include in datum_adapter.hpp - Remove anonymous namespace in server_status.cc - Remove incorrect comment in server_common.cc - Be rigorous about home threads in pseudo-table backends

+++ src/clustering/administration/datum_adapter.cc
@@ -46,7 +46,7 @@ bool convert_uuid_from_datum(
 }
 
 ql::datum_t convert_port_to_datum(
-        int value) {
+        uint16_t value) {
     return ql::datum_t(static_cast<double>(value));
 }
